TST: show reason for failure in test suite.
I am in a bad mood. I tried for a very long time to figure out why
this test fails on windows. There is very inconsistent behavior related
to inspecting the execution of these async methods and I never want to
debug this function ever again. I didn't even NEED to figure this out,
but it was upsetting me and then I kept going and it got worse and worse
and it was a bad time and I'm not going to ever be happy about this
until I understand why the test fails on windows and also why the output
I was seeing was inconsistent and why stdout and logging and raising
exceptions don't work as I expect. I was especially weirded out when one
iteration of the test showed a NotImplementedError but it never came
back again.